advanced-api-project/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-filter==24.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2 